KTO-826: Refactor components for KOMO tutkinnon osa
- Move TutkinnonOsienKuvausSection, TutkinnonOsatSection, useLocalizedKoulutus and InfoBoxGrid (renamed) to separate files from KoulutuksenTiedotSection and KoulutusForm/TiedotSection
